set: optimize lookup in compacted set
Instead of retrying the lookup from the root at each resolution when we
have a not found, walk the automata and check if we're in a final state
at each step.

    group                 v1                          v2
    -----                 --                          --
    Lookup/Compacted/0    1.00    146.9±0.83ns        1.09    160.7±0.86ns
    Lookup/Compacted/1    1.86    280.2±1.99ns        1.00    150.5±0.79ns
    Lookup/Compacted/2    2.72    341.1±2.43ns        1.00    125.2±0.70ns
    Lookup/Compacted/3    3.69    443.3±3.19ns        1.00    120.3±0.78ns
    Lookup/Compacted/4    4.57    532.0±3.00ns        1.00    116.3±0.47ns
    Lookup/Compacted/5    6.06    523.1±3.12ns        1.00     86.3±0.48ns
    Lookup/Compacted/6    6.81    572.5±3.48ns        1.00     84.1±0.59ns
    Lookup/Compacted/7    7.69    535.0±3.20ns        1.00     69.5±0.34ns

We do have a little slowdown for non-compressed key, i.e. when we need
to match the whole key, because we now check at every step instead of
once at the end.
On the other hand, lookup in compacted set are now faster than in
uncompacted one (which is pretty nice since compacted set are built way
faster), and the more compacted the set is, the faster the lookup!